refactor: Organize services by domain - books module
Clean Architecture Organization:
- Reorganized all book-related services into dedicated folder structure
- Created barrel exports for clean imports
- Fixed all relative import paths after reorganization

Services Reorganized:
1. BookImageService → src/infrastructure/services/books/
2. BookPDFService → src/infrastructure/services/books/
3. PDFExtractorService → src/infrastructure/services/books/
4. BookReadingAnalyticsService → src/infrastructure/services/books/ (renamed from BookReadingAnalytics.service.ts)

New Structure:
src/infrastructure/services/
├── books/
│   ├── BookImageService.ts
│   ├── BookPDFService.ts
│   ├── PDFExtractorService.ts
│   ├── BookReadingAnalyticsService.ts
│   └── index.ts (barrel export)
├── i18n/
│   ├── TranslationService.ts
│   └── index.ts
└── index.ts (main barrel export)

Updated Imports in:
- app/[locale]/books/[id]/read/page.tsx
- app/[locale]/books/[id]/statistics/page.tsx
- src/presentation/features/books/hooks/useBookForm.ts
- src/presentation/features/books/hooks/useReadingAnalytics.ts

Fixed:
- Relative import paths in BookReadingAnalyticsService (../ → ../../)
- Dynamic imports for PDFExtractorService in read page and hooks
- All imports now use clean barrel exports: @/src/infrastructure/services/books

Benefits:
- Easy to find book-related services
- Scalable structure for future modules
- Clean imports without deep paths
- Consistent with repository organization

Result: Build successful ✅
