+++ src/infrastructure/services/books/BookImageService.ts
