+++ src/infrastructure/services/books/BookPDFService.ts
