+++ src/infrastructure/services/books/PDFExtractorService.ts
