+++ src/infrastructure/services/books/index.ts
@@ -1,0 +1,9 @@
+// ============================================
+// src/infrastructure/services/books/index.ts
+// Barrel export for book-related services
+// ============================================
+
+export { BookImageService } from './BookImageService';
+export { BookPDFService } from './BookPDFService';
+export { PDFExtractorService } from './PDFExtractorService';
+export { BookReadingAnalyticsService } from './BookReadingAnalyticsService';
