+++ src/infrastructure/services/index.ts
@@ -1,0 +1,10 @@
+// ============================================
+// src/infrastructure/services/index.ts
+// Barrel export for all infrastructure services
+// ============================================
+
+// Books services
+export * from './books';
+
+// Internationalization services
+export * from './i18n';
